+++ front/src/App.js
@@ -1,13 +1,17 @@
-import React from 'react';
+import React, {useState} from 'react';
 import './styles/tailwind.css';
 import FrontPage from "./Pages/FrontPage";
 import General from "./Layout/General";
-import Login from "./Pages/Login";
+import Login from "./Pages/Login/Login";
 import Results from "./Pages/Results";
 import {BrowserRouter as Router, Switch, Route} from "react-router-dom";
 import "./app.scss"
+import { ApolloProvider } from '@apollo/client';
+import {client} from "./client";
 function App() {
-    return <div className="App">
+
+    return <ApolloProvider client={client}>
+        <div className="App">
         <Router>
             <General>
                 <Switch>
@@ -17,7 +21,8 @@ function App() {
                 </Switch>
             </General>
         </Router>
-    </div>;
+    </div>
+    </ApolloProvider>;
 }
 
 export default App;
