+++ front/src/Components/Icons/Icons.js
@@ -3,7 +3,7 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome"
 import {faMoneyBillAlt} from "@fortawesome/free-regular-svg-icons";
 
 import {faCreditCard} from "@fortawesome/free-regular-svg-icons";
-import { faCloudUploadAlt, faCloudDownloadAlt, faSignInAlt, faPiggyBank, faCalculator, faExclamationCircle, faTimesCircle, faInfoCircle} from "@fortawesome/free-solid-svg-icons";
+import { faCloudUploadAlt, faCloudDownloadAlt, faSignInAlt, faPiggyBank, faCalculator, faExclamationCircle, faTimesCircle, faInfoCircle, faCheckCircle} from "@fortawesome/free-solid-svg-icons";
 import {faFacebookSquare, faGoogle, faApple} from "@fortawesome/free-brands-svg-icons";
 
 export const UploadCloud = () => <FontAwesomeIcon icon={faCloudUploadAlt}/>
@@ -19,3 +19,4 @@ export const MoneyBill = () => <FontAwesomeIcon icon={faMoneyBillAlt}/>
 export const ErrorMark = () => <FontAwesomeIcon icon={faTimesCircle}/>
 export const WarningMark = () => <FontAwesomeIcon icon={faExclamationCircle}/>
 export const InfoMark = () => <FontAwesomeIcon icon={faInfoCircle}/>
+export const Ok = () => <FontAwesomeIcon icon={faCheckCircle}/>
