+++ front/src/Components/Messages/Message.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {ErrorMark, InfoMark, WarningMark} from "../Icons/Icons";
+import {ErrorMark, InfoMark, Ok, WarningMark} from "../Icons/Icons";
 
 export const BaseMessage = ({color, icon, message, children}) => <div className={`
     pb-2 pt-2
@@ -21,5 +21,6 @@ export const BaseMessage = ({color, icon, message, children}) => <div className=
 export const Error = ({message, children}) => <BaseMessage color="red" icon={<ErrorMark />} message={message}>{children}</BaseMessage>
 export const Notice = ({message, children}) => <BaseMessage color="yellow" icon={<WarningMark />} message={message}>{children}</BaseMessage>
 export const Info = ({message, children}) => <BaseMessage color="blue" icon={<InfoMark />} message={message}>{children}</BaseMessage>
+export const Success = ({message, children}) => <BaseMessage color="green" icon={<Ok />} message={message}>{children}</BaseMessage>
